Backend/requirements.txt
absl-py==2.2.1
annotated-types==0.7.0
asgiref==3.8.1
astunparse==1.6.3
cachetools==5.5.2
certifi==2025.1.31
charset-normalizer==3.4.1
colorama==0.4.6
Django==5.1.7
django-cors-headers==4.7.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
dotenv==0.9.9
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.3.2
gast==0.6.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.181.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
google-generativeai==0.8.5
google-pasta==0.2.0
googleapis-common-protos==1.70.0
grpcio==1.74.0
grpcio-status==1.71.2
h5py==3.13.0
httplib2==0.31.0
idna==3.10
Jinja2==3.1.6
keras==3.9.2
Keras-Preprocessing==1.1.2
libclang==18.1.1
Markdown==3.7
markdown-it-py==3.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
ml_dtypes==0.5.1
mpmath==1.3.0
namex==0.0.8
networkx==3.4.2
numpy==2.1.3
oauthlib==3.2.2
opt_einsum==3.4.0
optree==0.14.1
packaging==24.2
pillow==11.1.0
proto-plus==1.26.1
protobuf==5.29.4
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.1
PyJWT==2.9.0
pyparsing==3.2.3
python-dotenv==1.1.1
requests==2.32.3
requests-oauthlib==2.0.0
rich==14.0.0
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
tensorboard==2.19.0
tensorboard-data-server==0.7.2
tensorboard-plugin-wit==1.8.1
tensorflow==2.19.0
tensorflow-estimator==2.10.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==3.0.1
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.13.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.3.0
Werkzeug==3.1.3
wrapt==1.17.2